backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
langchain>=0.1.16
langgraph>=0.0.38
langchain-openai>=0.1.3
httpx>=0.27.0
